🌱 Split UpdatesTabRequestItem into co-located sub-components
Extract ExternalLinkRow, the triaged/untriaged content blocks, the
fix-complete banner and verification prompt, the preview section and the
request actions row into sibling modules. UpdatesTabRequestItem.tsx goes
from 648 to 219 lines / 5 hooks with no behaviour change.
